only strip quotes from name taken from address
phpMailer is adding quotes to the name field, and escaping quotes which are already in the name field.  So they get removed in cleanAddress() to prevent quoted quotes within quotes.

I think there should be a single place where names/addresses are validated and formatted, so that the behaviour is more predictable, or so that it's easier to track down why it behaves oddly. ;-)

Right now if I pass in 'Bobby Fisher (World Champion) <bobby@wcf.org>' then the header will get '"Bobby Fisher (World Champion)" <bobby@wcf.org>"

But if I pass 'Bobby Fisher <bobby@wcf.org>' then the header gets it verbatim.

Where this quoting happens is not immediately apparent, but it makes a mess of quotes if I leave them in the input, and adds them /sometimes/ in the output, which is a bit frustrating if you want consistent behaviour.