Also use validateAddress in addOrEnqueueAnAddress
Prior to this commit E-Mails are if they contain an `@` sign in
`PHPMailer\addOrEnqueueAnAddress` before reaching the configurable
`validateAddress` in `PHPMailer\addAnAddress`. In 99.9999% of the
cases this will be perfectly fine. But however, it is possible to send
mail only to a user on the local system by their username. This is
especially neat for local development where you just have a sendmail
delivering mails to local user. Depending on your sendmail backend, it
would require additional configuration to create aliases with a domain
(e.g. something with `@something.tld`).

And if you provide a custom validator for that, I don't see any reason
to not use the configurable filter in that scenario.