Composer: update some CS related dependencies
The Composer PHPCS plugin has released its 1.0.0 version. :tada:

As minors below a 1.0 are considered as majors by Composer, this needs an update to the `composer.json` file to get access to the new version.
And as PHPCS has also released a new version, may as well update that as well.

Refs:
* https://github.com/PHPCSStandards/composer-installer/releases/tag/v1.0.0
* https://github.com/squizlabs/PHP_CodeSniffer/releases/tag/3.7.2